requirements.txt
argon2-cffi==16.1.0
cffi==1.7.0
cryptography==1.4
google-api-python-client==1.5.1
httplib2==0.9.2
idna==2.1
oauth2client==3.0.0
otpauth==1.0.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
PyDrive3==1.0.0
PyNaCl==1.0.1
pyOpenSSL==16.0.0
PyYAML==3.11
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
